tests: Move SemaphoreExportImportSupported to file
