+++ tests/framework/layer_validation_tests.cpp
@@ -376,17 +376,6 @@ VkFormat FindFormatWithoutFeatures2(VkPhysicalDevice gpu, VkImageTiling tiling, 
     return return_format;
 }
 
-bool SemaphoreExportImportSupported(VkPhysicalDevice gpu, VkExternalSemaphoreHandleTypeFlagBits handle_type) {
-    constexpr auto export_import_flags =
-        VK_EXTERNAL_SEMAPHORE_FEATURE_EXPORTABLE_BIT_KHR | VK_EXTERNAL_SEMAPHORE_FEATURE_IMPORTABLE_BIT_KHR;
-
-    VkPhysicalDeviceExternalSemaphoreInfo info = vku::InitStructHelper();
-    info.handleType = handle_type;
-    VkExternalSemaphoreProperties properties = vku::InitStructHelper();
-    vk::GetPhysicalDeviceExternalSemaphoreProperties(gpu, &info, &properties);
-    return (properties.externalSemaphoreFeatures & export_import_flags) == export_import_flags;
-}
-
 void CreateSamplerTest(VkLayerTest &test, const VkSamplerCreateInfo *create_info, const std::string &code) {
     if (code.length()) {
         test.Monitor().SetDesiredFailureMsg(kErrorBit, code);
